+++ passes/testdata/src/github.com/hashicorp/terraform-plugin-sdk/v2/helper/schema/schema.go
@@ -41,6 +41,7 @@ type Schema struct {
 	ForceNew      bool
 	Sensitive     bool
 	Default       interface{}
+	Description   string
 	MaxItems      int
 	Elem          interface{}
 	ValidateFunc  SchemaValidateFunc