+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## v0.2.4 (2026-05-xx)
+
+- New rule `AZSD005`: check that `None` enum values are not listed in `StringInSlice` validation
+- New rule `AZNR010`: check that schema fields do not have redundant `Description`
+
+
 ## v0.2.3 (2026-05-12)
 
 - New rule `AZNR009`: check that API response IDs are parsed before assigning to state